pyproject.toml
[project]
name = "my-app"
version = "0.1.0"
description = ""
authors = [
    {name = "mikihiromakino",email = "16id014mmiki@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
